+++ src/sqlfluff/dialects/dialect_mysql.py
@@ -17,6 +17,8 @@ from sqlfluff.core.parser import (
     CodeSegment,
     StringParser,
     SymbolSegment,
+    Anything,
+    Delimited,
 )
 from sqlfluff.core.dialects import load_raw_dialect
 
@@ -102,6 +104,28 @@ mysql_dialect.add(
     DoubleDollarSignSegment=StringParser(
         "$$", SymbolSegment, name="doubledollarsign", type="statement_terminator"
     ),
+    AmpersandSignSegment=StringParser(
+        "@", SymbolSegment, name="ampersandsign", type="user_designator"
+    ),
+    OutputParameterSegment=StringParser(
+        "OUT", SymbolSegment, name="inputparameter", type="parameter_direction"
+    ),
+    InputParameterSegment=StringParser(
+        "IN", SymbolSegment, name="outputparameter", type="parameter_direction"
+    ),
+    InputOutputParameterSegment=StringParser(
+        "INOUT", SymbolSegment, name="inputoutputparameter", type="parameter_direction"
+    ),
+    ProcedureParameterGrammar=OneOf(
+        Sequence(
+            OneOf(Ref("OutputParameterSegment"),
+                  Ref("InputParameterSegment"),
+                  Ref("InputOutputParameterSegment"), optional=True),
+            Ref("ParameterNameSegment", optional=True),
+            Ref("DatatypeSegment"),
+        ),
+        Ref("DatatypeSegment"),
+    ),
 )
 
 mysql_dialect.replace(
@@ -120,10 +144,124 @@ class DelimiterStatement(BaseSegment):
     match_grammar = Ref.keyword("DELIMITER")
 
 
+@mysql_dialect.segment()
+class CreateProcedureStatementSegment(BaseSegment):
+    """A `CREATE PROCEDURE` statement.
+
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/create-procedure.html
+    """
+
+    type = "create_procedure_statement"
+
+    match_grammar = Sequence(
+        "CREATE",
+        "PROCEDURE",
+        Anything(),
+    )
+
+    parse_grammar = Sequence(
+        "CREATE",
+        "PROCEDURE",
+        Ref("FunctionNameSegment"),
+        Ref("ProcedureParameterListGrammar", optional=True),
+        Ref("CommentClauseSegment", optional=True),
+        Ref("CharacteristicStatement", optional=True),
+        Ref("FunctionDefinitionGrammar"),
+    )
+
+
 @mysql_dialect.segment(replace=True)
 class StatementSegment(ansi_dialect.get_segment("StatementSegment")):  # type: ignore
-    """Overriding StatementSegment to allow for DELIMITER."""
+    """Overriding StatementSegment to allow for additional segment parsing."""
 
     parse_grammar = ansi_dialect.get_segment("StatementSegment").parse_grammar.copy(
-        insert=[Ref("DelimiterStatement")],
+        insert=[
+            Ref("DelimiterStatement"),
+            Ref("CreateProcedureStatementSegment"),
+        ],
+    )
+
+
+@mysql_dialect.segment(replace=True)
+class FunctionDefinitionGrammar(BaseSegment):
+    """This is the body of a `CREATE FUNCTION` statement."""
+
+    match_grammar = Sequence(
+        Ref("TransactionStatementSegment"),
+        Ref("StatementSegment"),
+    )
+
+
+@mysql_dialect.segment()
+class CharacteristicStatement(BaseSegment):
+    type = "characteristic_statement"
+    match_grammar = Sequence(
+        OneOf("DETERMINISTIC", Sequence("NOT", "DETERMINISTIC")),
+        Sequence("LANGUAGE", "SQL", optional=True),
+        OneOf(Sequence("CONTAINS", "SQL", optional=True),
+              Sequence("NO", "SQL", optional=True),
+              Sequence("READS", "SQL", "DATA", optional=True),
+              Sequence("MODIFIES", "SQL", "DATA", optional=True), optional=True),
+        Sequence("SQL", "SECURITY", OneOf("DEFINER", "INVOKER"), optional=True),
+    )
+
+
+@mysql_dialect.segment(replace=True)
+class CreateFunctionStatementSegment(BaseSegment):
+    """A `CREATE FUNCTION` statement.
+
+    mysql: https://dev.mysql.com/doc/refman/8.0/en/create-procedure.html
+    """
+
+    type = "create_function_statement"
+
+    match_grammar = Sequence(
+        "CREATE",
+        "FUNCTION",
+        Anything(),
+    )
+
+    parse_grammar = Sequence(
+        "CREATE",
+        "FUNCTION",
+        Ref("FunctionNameSegment"),
+        Ref("FunctionParameterListGrammar", optional=True),
+        Sequence("RETURNS", Ref("DatatypeSegment"),),
+        Ref("CommentClauseSegment", optional=True),
+        Ref("CharacteristicStatement"),
+        Ref("FunctionDefinitionGrammar"),
+    )
+
+
+@mysql_dialect.segment(replace=True)
+class DropStatementSegment(BaseSegment):
+    """A `DROP` statement."""
+
+    type = "drop_statement"
+
+    match_grammar = Sequence(
+        "DROP",
+        OneOf(
+            "TABLE",
+            "VIEW",
+            "USER",
+            "FUNCTION",
+            "PROCEDURE",
+        ),
+        Ref("IfExistsGrammar", optional=True),
+        Ref("TableReferenceSegment"),
+    )
+
+
+@mysql_dialect.segment()
+class ProcedureParameterListGrammar(BaseSegment):
+    """The parameters for a procedure ie. `(string, number)`."""
+
+    # Function parameter list
+    match_grammar = Bracketed(
+        Delimited(
+            Ref("ProcedureParameterGrammar"),
+            delimiter=Ref("CommaSegment"),
+            optional=True,
+        ),
     )