+++ test/fixtures/parser/mysql/procedure_inout_param.sql
@@ -1,0 +1,3 @@
+CREATE PROCEDURE `testprocedure`(inout test int)
+BEGIN
+END~