+++ test/fixtures/parser/mysql/procedure_out_param.sql
@@ -1,0 +1,3 @@
+CREATE PROCEDURE `testprocedure`(out test int)
+BEGIN
+END~